+++ src/realm/table.hpp
@@ -1176,6 +1176,9 @@ private:
 
     //@}
 
+    /// Used by query. Follows chain of link columns and returns final target table
+    const Table* get_link_chain_target(const std::vector<size_t>& link_chain) const;
+
     /// Remove the specified row by the 'move last over' method.
     void do_move_last_over(std::size_t row_ndx);
 
@@ -1608,6 +1611,24 @@ template<class T> inline Columns<T> Table::column(std::size_t column)
     if (std::is_same<T, Link>::value || std::is_same<T, LinkList>::value) {
         tmp.push_back(column);
     }
+
+    // Check if user-given template type equals Realm type. Todo, we should clean up and reuse all our 
+    // type traits (all the is_same() cases below).
+    const Table* table = get_link_chain_target(m_link_chain);
+
+    realm::DataType ct = table->get_column_type(column);
+    if (std::is_same<T, int64_t>::value && ct != type_Int)
+        throw(LogicError::type_mismatch);
+    else if (std::is_same<T, bool>::value && ct != type_Bool)
+        throw(LogicError::type_mismatch);
+    else if (std::is_same<T, realm::DataType>::value && ct != type_DateTime)
+        throw(LogicError::type_mismatch);
+    else if (std::is_same<T, float>::value && ct != type_Float)
+        throw(LogicError::type_mismatch);
+    else if (std::is_same<T, double>::value && ct != type_Double)
+        throw(LogicError::type_mismatch);
+
+
     m_link_chain.clear();
     return Columns<T>(column, this, tmp);
 }
@@ -2199,6 +2220,11 @@ public:
     static bool is_cross_table_link_target(const Table& table)
     {
         return table.is_cross_table_link_target();
+    }
+
+    static Group* get_parent_group(const Table& table) REALM_NOEXCEPT
+    {
+        return table.get_parent_group();
     }
 
     static Replication* get_repl(Table& table) REALM_NOEXCEPT